fix mllm missing issue
